packages/v1/external_accounts/requirements.txt
cloudinary==1.39.0
google-cloud-secret-manager==2.20.2
openai==1.36.0
psycopg2-binary==2.9.9
PyJWT==1.7.1
python-dotenv==1.0.1
requests==2.31.0
SQLAlchemy==2.0.28
stripe==5.0.0
urllib3==1.26.19
Werkzeug==3.0.3
